remote_execution: implement message debugging
This commit add serde to re_grpc_proto package so that we could convert
request/response message to JSON for debugging purposes.

Added debugging methods so _SOME_ RPC calls to demonstrate the setup.

The ByteStream and GetCapabilities RPC remains unimplemented and shall be
left as an excercise for our readers.